requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
requests==2.31.0
pydantic>=2.0.0
anthropic>=0.40.0
python-dotenv==1.0.0